Move the palette logic from Shell to plugin
